skud/requirements.txt
mfrc522==1.0.3      # драйвер для RC522 (SPI)
PyYAML>=6.0         # для загрузки config.yaml
requests>=2.28.0    # HTTP-клиент для отправки POST-запросов